refactor code for reuse
