+++ src/main/java/bio/terra/service/snapshotbuilder/ConceptChildrenQueryBuilder.java
@@ -1,4 +1,4 @@
-package bio.terra.service.snapshotbuilder.utils;
+package bio.terra.service.snapshotbuilder;
 
 import bio.terra.service.snapshotbuilder.query.FieldPointer;
 import bio.terra.service.snapshotbuilder.query.FieldVariable;
